% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{DetectionsResource}
\alias{DetectionsResource}
\title{DetectionsResource Object}
\usage{
DetectionsResource()
}
\value{
DetectionsResource object
}
\description{
DetectionsResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{DetectionsListResponse}},
  \code{\link{LanguagesListResponse}},
  \code{\link{LanguagesResource}},
  \code{\link{TranslationsListResponse}},
  \code{\link{TranslationsResource}}
}

