% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{language.detections.list}
\alias{language.detections.list}
\title{Detect the language of text.}
\usage{
language.detections.list(q)
}
\arguments{
\item{q}{The text to detect}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/translate/v2/using_rest}{Google Documentation}
}

