% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{language.languages.list}
\alias{language.languages.list}
\title{List the source/target languages supported by the API}
\usage{
language.languages.list(target = NULL)
}
\arguments{
\item{target}{the language and collation in which the localized results should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/translate/v2/using_rest}{Google Documentation}
}

