% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\name{language.translations.list}
\alias{language.translations.list}
\title{Returns text translations from one language to another.}
\usage{
language.translations.list(q, target, cid = NULL, format = NULL,
  source = NULL)
}
\arguments{
\item{q}{The text to translate}

\item{target}{The target language into which the text should be translated}

\item{cid}{The customization id for translate}

\item{format}{The format of the text}

\item{source}{The source language of the text}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/translate/v2/using_rest}{Google Documentation}
}

