% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{LanguagesListResponse}
\alias{LanguagesListResponse}
\title{LanguagesListResponse Object}
\usage{
LanguagesListResponse(languages = NULL)
}
\arguments{
\item{languages}{List of source/target languages supported by the translation API}
}
\value{
LanguagesListResponse object
}
\description{
LanguagesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
