% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{LanguagesResource}
\alias{LanguagesResource}
\title{LanguagesResource Object}
\usage{
LanguagesResource(language = NULL, name = NULL)
}
\arguments{
\item{language}{The language code}

\item{name}{The localized name of the language if target parameter is given}
}
\value{
LanguagesResource object
}
\description{
LanguagesResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{DetectionsListResponse}},
  \code{\link{DetectionsResource}},
  \code{\link{LanguagesListResponse}},
  \code{\link{TranslationsListResponse}},
  \code{\link{TranslationsResource}}
}

