% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{TranslationsListResponse}
\alias{TranslationsListResponse}
\title{TranslationsListResponse Object}
\usage{
TranslationsListResponse(translations = NULL)
}
\arguments{
\item{translations}{Translations contains list of translation results of given text}
}
\value{
TranslationsListResponse object
}
\description{
TranslationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{DetectionsListResponse}},
  \code{\link{DetectionsResource}},
  \code{\link{LanguagesListResponse}},
  \code{\link{LanguagesResource}},
  \code{\link{TranslationsResource}}
}

