% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{TranslationsResource}
\alias{TranslationsResource}
\title{TranslationsResource Object}
\usage{
TranslationsResource(detectedSourceLanguage = NULL, translatedText = NULL)
}
\arguments{
\item{detectedSourceLanguage}{Detected source language if source parameter is unspecified}

\item{translatedText}{The translation}
}
\value{
TranslationsResource object
}
\description{
TranslationsResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{DetectionsListResponse}},
  \code{\link{DetectionsResource}},
  \code{\link{LanguagesListResponse}},
  \code{\link{LanguagesResource}},
  \code{\link{TranslationsListResponse}}
}

