% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_objects.R
\name{DetectionsListResponse}
\alias{DetectionsListResponse}
\title{Translate API Objects 
Translates text from one language to another.}
\usage{
DetectionsListResponse(detections = NULL)
}
\arguments{
\item{detections}{A detections contains detection results of several text}
}
\value{
DetectionsListResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:50:08
filename: /Users/mark/dev/R/autoGoogleAPI/googletranslatev2.auto/R/translate_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DetectionsListResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{DetectionsResource}},
  \code{\link{LanguagesListResponse}},
  \code{\link{LanguagesResource}},
  \code{\link{TranslationsListResponse}},
  \code{\link{TranslationsResource}}
}

