% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_functions.R
\docType{package}
\name{translate_googleAuthR}
\alias{translate_googleAuthR}
\alias{translate_googleAuthR-package}
\title{Translate API
Translates text from one language to another.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:50:08
filename: /Users/mark/dev/R/autoGoogleAPI/googletranslatev2.auto/R/translate_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

